% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_beta_est.R
\name{opt_beta_est}
\alias{opt_beta_est}
\title{Beta optimization function}
\usage{
opt_beta_est(X, y)
}
\arguments{
\item{X}{This is a matrix of data with dimensions n x p.}

\item{y}{This is a vector with a length of n.}
}
\value{
This function will return an estimate of the parameter beta, along with an intercept.
}
\description{
This function calculates the optimal values for the parameter beta when provided with a matrix of data called X, and
the the corresponding y binary response vector. When these values are input, the function will find an initial beta
beta estimate, then uses an optimization function to refine the beta estimates until they are sufficiently near
the true values of beta.
}
\examples{
n <- 100
p <- 3
p_i <- 1 / (1 + exp(-design \%*\% beta_true))
y <- rbinom(n, size = 1, prob = p_i)
beta_opt <- opt_beta_est(X = X, y = y)
print(beta_opt)
}
